[mono] Workaround for the assumption that `dotnet` is installed
We moved to building purely with mono, earlier and thus skip the dotnet
installation in the scripts. This breaks that. Fixing it up in-place for
now.

Commit 2: (Patch on behalf of Alexander Köplinger)
Also, fix the command line, as the `dotnet nuget` uses a different
command line syntax than the standalone `nuget` that mono bundles.

    `nuget locals http-cache -c` -> `nuget locals http-cache -Clear`