Add support for .NET Core 3.1 frameworks
Most importantly, allows the xplat msbuild to build netcoreapp3.1 projects.

Updates the Microsoft.NETCoreSdk.BundledVersions.props file. This is the
output of GenerateBundledVersions.targets, which is a part of the sdk installer.

It seems that this file was updated manually in the past. Implicit
package versions are only used for versions <3.0.

I have kept the
<NETCoreSdkRuntimeIdentifier>osx-x64</NETCoreSdkRuntimeIdentifier>,
but I believe this is wrong, as this file is not only used on osx
platforms. I have absolutely no idea what happens if you try to use
this on Windows, it does, however, work on Linux.

RuntimeIdentifierGraph.json being hardcoded is also not great, as there
are installations (such as the version distributed by Arch Linux), that
use platforms that are patched in (arch-x64), and as such, they are not
in this one list.

Either way, this commit does not worsen the situation in any way.

Ultimately, these files should probably be taken from the dotnet sdk
installation and not hardcoded here in this way. I do not know enough to
see if it's feasible, or if this is as good of a solution as we can
have.